+++ plex/plextv/auth_tokensource_test.go
@@ -1,16 +1,18 @@
-package plexauth
+package plextv
 
 import (
 	"context"
 	"errors"
 	"log/slog"
 	"os"
+	"reflect"
+	"sync/atomic"
 	"testing"
 	"time"
 )
 
 func TestTokenSource_WithToken(t *testing.T) {
-	ts := DefaultConfig.TokenSource(WithToken("abc"))
+	ts := DefaultConfig().TokenSource(WithToken("abc"))
 	token, err := ts.Token(t.Context())
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
@@ -22,80 +24,59 @@ func TestTokenSource_WithToken(t *testing.T) {
 
 func TestTokenSource_WithCredentials(t *testing.T) {
 	// auth server
-	cfg, _, s := newTestServer(DefaultConfig.WithClientID("my-client-id"))
+	cfg, _, s := newTestServer(DefaultConfig().WithClientID("my-client-id"))
 	t.Cleanup(s.Close)
-
-	// happy path
-	ts := cfg.TokenSource(WithCredentials("user", "pass"))
-	token, err := ts.Token(t.Context())
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if token.String() != legacyToken {
-		t.Fatalf("unexpected token: %s", token)
-	}
-
-	// clear the cached token
-	ts.(*cachingTokenSource).token = nil
-	// a failed registrar will fail the token source
-	ts.(*cachingTokenSource).tokenSource = fakeRegistrar{err: errors.New("test error")}
-	_, err = ts.Token(t.Context())
-	if err == nil {
-		t.Fatalf("expected error, got nil")
-	}
+	doTokenSourceTest(t, cfg.TokenSource(WithCredentials("user", "pass")), legacyToken)
 }
 
 func TestTokenSource_WithPIN(t *testing.T) {
 	// auth server
-	cfg, _, s := newTestServer(DefaultConfig.WithClientID("my-client-id"))
+	cfg, _, s := newTestServer(DefaultConfig().WithClientID("my-client-id"))
 	t.Cleanup(s.Close)
-
-	// happy path
-	ts := cfg.TokenSource(WithPIN(func(_ PINResponse, _ string) {}, 100*time.Millisecond))
-	token, err := ts.Token(t.Context())
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if token.String() != legacyToken {
-		t.Fatalf("unexpected token: %s", token)
-	}
-
-	// clear the cached token
-	ts.(*cachingTokenSource).token = nil
-	// a failed registrar will fail the token source
-	ts.(*cachingTokenSource).tokenSource = fakeRegistrar{err: errors.New("test error")}
-	_, err = ts.Token(t.Context())
-	if err == nil {
-		t.Fatalf("expected error, got nil")
-	}
+	doTokenSourceTest(t, cfg.TokenSource(WithPIN(func(_ PINResponse, _ string) {}, 100*time.Millisecond)), legacyToken)
 }
 
 func TestTokenSource_WithJWT(t *testing.T) {
 	// auth server
-	cfg, s, tts := newTestServer(DefaultConfig.WithClientID("my-client-id"))
+	cfg, s, tts := newTestServer(DefaultConfig().WithClientID("my-client-id"))
 	t.Cleanup(tts.Close)
 
 	s.tokens.SetToken("my-client-id", legacyToken)
-
-	// happy path
 	var f fakeVault
 	ts := cfg.TokenSource(
 		WithCredentials("user", "pass"),
 		WithJWT(&f),
 		WithLogger(slog.New(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{Level: slog.LevelDebug}))),
 	)
+	doTokenSourceTest(t, ts, "")
+}
+
+func doTokenSourceTest(t *testing.T, ts TokenSource, want string) {
+	t.Helper()
+	// happy path
 	token, err := ts.Token(t.Context())
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
-	if !token.IsJWT() {
-		t.Fatal("expected JWT token")
+	if want != "" {
+		if got := token.String(); got != want {
+			t.Fatalf("unexpected token: want=%v, got=%v", want, token)
+		}
+	}
+
+	// clear the cached token
+	ts.(*cachingTokenSource).token = nil
+	// a failed registrar will fail the token source
+	ts.(*cachingTokenSource).tokenSource = fakeRegistrar{err: errors.New("test error")}
+	_, err = ts.Token(t.Context())
+	if err == nil {
+		t.Fatalf("expected error, got nil")
 	}
 }
 
 func Test_jwtTokenSource(t *testing.T) {
 	// auth server
-	cfg, s, tts := newTestServer(DefaultConfig.WithClientID("my-client-id"))
+	cfg, s, tts := newTestServer(DefaultConfig().WithClientID("my-client-id"))
 	t.Cleanup(tts.Close)
 
 	s.tokens.SetToken("my-client-id", legacyToken)
@@ -122,7 +103,7 @@ func Test_jwtTokenSource(t *testing.T) {
 
 func Test_jwtTokenSource_WrongClientID(t *testing.T) {
 	// auth server
-	cfg, s, tts := newTestServer(DefaultConfig.WithClientID("my-client-id"))
+	cfg, s, tts := newTestServer(DefaultConfig().WithClientID("my-client-id"))
 	t.Cleanup(tts.Close)
 
 	s.tokens.SetToken("my-client-id", legacyToken)
@@ -150,7 +131,7 @@ func Test_jwtTokenSource_WrongClientID(t *testing.T) {
 
 func TestTokenSource_NoRegistrar(t *testing.T) {
 	// auth server
-	cfg, _, s := newTestServer(DefaultConfig.WithClientID("my-client-id"))
+	cfg, _, s := newTestServer(DefaultConfig().WithClientID("my-client-id"))
 	s.Close()
 
 	var f fakeVault
@@ -176,6 +157,20 @@ func TestTokenSource_NoRegistrar(t *testing.T) {
 	}
 }
 
+func BenchmarkCachingTokenSource_Token_LegacyToken(b *testing.B) {
+	b.Run("legacy", func(b *testing.B) {
+		ts := cachingTokenSource{tokenSource: fakeRegistrar{token: legacyToken}}
+		ctx := context.Background()
+		b.ReportAllocs()
+		for b.Loop() {
+			_, err := ts.Token(ctx)
+			if err != nil {
+				b.Fatal(err)
+			}
+		}
+	})
+}
+
 var _ TokenSource = fakeRegistrar{}
 
 type fakeRegistrar struct {
@@ -188,3 +183,57 @@ func (f fakeRegistrar) Token(_ context.Context) (Token, error) {
 }
 
 var _ secureDataVault = (*fakeVault)(nil)
+
+func TestJWTDataStore(t *testing.T) {
+	var v fakeVault
+	s := jwtDataStore{vault: &v, clientID: "my-client-id"}
+	if _, err := s.Load(); !errors.Is(err, os.ErrNotExist) {
+		t.Fatalf("expected os.ErrNotExist, got %v", err)
+	}
+
+	want := JWTSecureData{
+		KeyID:      "my-key-id",
+		ClientID:   "my-client-id",
+		PrivateKey: []byte("my-private-key"),
+	}
+	if err := s.Save(want); err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+	got, err := s.Load()
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+	if !reflect.DeepEqual(got, want) {
+		t.Fatalf("Load() want: %+v, got: %+v", want, got)
+	}
+
+	// store the wrong client ID
+	data, _ := v.Load()
+	data.ClientID = "invalid-client-id"
+	_ = v.Save(data)
+
+	got, err = s.Load()
+	if !errors.Is(err, ErrInvalidClientID) {
+		t.Fatalf("expected error, got nil")
+	}
+}
+
+type fakeVault struct {
+	data atomic.Pointer[JWTSecureData]
+	err  error
+}
+
+func (f *fakeVault) Load() (JWTSecureData, error) {
+	if f.err != nil {
+		return JWTSecureData{}, f.err
+	}
+	if data := f.data.Load(); data != nil {
+		return *data, nil
+	}
+	return JWTSecureData{}, os.ErrNotExist
+}
+
+func (f *fakeVault) Save(data JWTSecureData) error {
+	f.data.Store(&data)
+	return nil
+}
