+++ plex/plextv/errors.go
@@ -1,4 +1,4 @@
-package plexauth
+package plextv
 
 import (
 	"bytes"
@@ -12,13 +12,24 @@ import (
 var (
 	// ErrInvalidToken indicates that a Config method was passed an invalid token.
 	ErrInvalidToken = fmt.Errorf("invalid token")
+	// ErrInvalidClientID indicates that the data in a JWTSecureDataStore is for a different Client ID
+	// than the one used to register the device.
+	// The stored private key and public key ID can't be used to generate a new JWT token.
+	// You will need to re-register the device with a new Client ID.
+	ErrInvalidClientID = fmt.Errorf("data store contains invalid client ID")
 	// ErrNoTokenSource indicates that a token source needs a child token source, but none was provided.
 	// A typical example is JWTTokenSource or PMSTokenSource needing a registrar to get a legacy token,
 	// but none is provided in [Config.TokenSource].
-	ErrNoTokenSource   = errors.New("no token source provided")
-	ErrUnauthorized    = errors.New("user could not be authenticated")
+	ErrNoTokenSource = errors.New("no token source provided")
+	// ErrUnauthorized indicates that plex.tv could not authenticate the user.
+	ErrUnauthorized = errors.New("user could not be authenticated")
+	// ErrTooManyRequests indicates that the plex.tv API rate limit has been reached.
+	// Typically, this only happens when generating a new JWT token.
 	ErrTooManyRequests = errors.New("too many requests")
-	ErrJWKMissing      = errors.New("jwk missing. no public key to verify jwt request")
+	// ErrJWKMissing indicates that the JWT token request could not find a public key to verify the JWT request.
+	// Typically, this means the Registered Device for the associated Client ID has been removed from Plex's Registered Devices.
+	// You will need to re-register the device with a new Client ID and a new public/private key pair.
+	ErrJWKMissing = errors.New("jwk missing. no public key to verify jwt request")
 )
 
 var _ error = &PlexError{}
