+++ plex/plextv/errors_test.go
@@ -1,4 +1,4 @@
-package plexauth
+package plextv
 
 import (
 	"bytes"
