+++ plex/plextv/plextv_types.go
@@ -1,4 +1,4 @@
-package plexauth
+package plextv
 
 import (
 	"encoding/xml"
