+++ plex/plextv/plextv_types_test.go
@@ -1,4 +1,4 @@
-package plexauth
+package plextv
 
 import (
 	"encoding/xml"
