+++ plex/plextv/token.go
@@ -1,4 +1,4 @@
-package plexauth
+package plextv
 
 import (
 	"errors"
@@ -19,42 +19,51 @@ const clockSkewTolerance = time.Minute
 type Token string
 
 // String returns the token as a string.
-func (t Token) String() string {
-	return string(t)
+func (t *Token) String() string {
+	if t == nil {
+		return ""
+	}
+	return string(*t)
 }
 
 // IsLegacy returns true if the token is a legacy token (20-character string).
-func (t Token) IsLegacy() bool {
-	return len(t) == 20
+func (t *Token) IsLegacy() bool {
+	if t == nil {
+		return false
+	}
+	return len(*t) == 20
 }
 
 // IsJWT returns true if the token is a JWT.
 // Note: returns true even if the token is invalid (e.g. expired).
-func (t Token) IsJWT() bool {
+func (t *Token) IsJWT() bool {
+	if t == nil {
+		return false
+	}
 	_, err := t.parseJWT()
 	return err == nil || errors.Is(err, jwt.TokenExpiredError())
 }
 
 // IsValid returns true if the token is valid.
 // Note: for JWT, the token is valid if it is not expired. The signature, if present, is not verified.
-func (t Token) IsValid() bool {
+func (t *Token) IsValid() bool {
 	// parsing each time we validate the token isn't very efficient. currently not an issue as jwt tokens
 	// are not maintained (only needed to get the pms token, which is then cached), but may become an issue.
 	// TODO: PlexTVClient may invalidate that premise.
-	if t.IsLegacy() {
-		return t != ""
+	if t == nil {
+		return false
 	}
 	tok, err := t.parseJWT()
-	if err != nil {
+	if err == nil {
+		exp, ok := tok.Expiration()
+		if ok && exp.After(time.Now()) {
+			return true
+		}
 		return false
 	}
-	exp, ok := tok.Expiration()
-	if !ok {
-		return false
-	}
-	return exp.After(time.Now())
+	return t.IsLegacy()
 }
 
-func (t Token) parseJWT() (jwt.Token, error) {
-	return jwt.Parse([]byte(t), jwt.WithVerify(false), jwt.WithAcceptableSkew(clockSkewTolerance))
+func (t *Token) parseJWT() (jwt.Token, error) {
+	return jwt.Parse([]byte(*t), jwt.WithVerify(false), jwt.WithAcceptableSkew(clockSkewTolerance))
 }
