+++ plex/plextv/until-successful.go
@@ -1,4 +1,4 @@
-package plexauth
+package plextv
 
 import (
 	"sync"
