+++ plex/plextv/until-successful_test.go
@@ -1,4 +1,4 @@
-package plexauth
+package plextv
 
 import (
 	"errors"
